Treat undefined properties as if they don't exist
This is an issue that arises when validating JS objects rather than objects
that have come straight from JSON. Properties that are undefined should be
treated as if they had undergone a JSON stringify/parse roundtrip: they should
be considered invisible when doing validation. This matches the behavior of
other JSON Schema validators that I have tested.